+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-java-sdk/src/main/java/com/microsoft/azure/toolkit/intellij/java/sdk/MavenProjectReportGenerator.java
@@ -21,8 +21,8 @@ import com.intellij.psi.xml.XmlTag;
 import com.microsoft.applicationinsights.TelemetryClient;
 import com.microsoft.applicationinsights.TelemetryConfiguration;
 import com.microsoft.applicationinsights.telemetry.EventTelemetry;
-import com.microsoft.azure.toolkit.intellij.java.sdk.report.*;
-import com.microsoft.azure.toolkit.intellij.java.sdk.report.Error;
+import com.microsoft.azure.toolkit.intellij.java.sdk.models.*;
+import com.microsoft.azure.toolkit.intellij.java.sdk.models.Error;
 import com.microsoft.azure.toolkit.intellij.java.sdk.utils.MavenUtils;
 import kotlin.Unit;
 import kotlin.coroutines.Continuation;
