+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-java-sdk/src/main/java/com/microsoft/azure/toolkit/intellij/java/sdk/models/MavenArtifactDetails.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.toolkit.intellij.java.sdk.utils;
+package com.microsoft.azure.toolkit.intellij.java.sdk.models;
 
 import java.time.OffsetDateTime;
