+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-java-sdk/src/main/java/com/microsoft/azure/toolkit/intellij/java/sdk/models/MethodCallDetails.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.azure.toolkit.intellij.java.sdk.report;
+package com.microsoft.azure.toolkit.intellij.java.sdk.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
