+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-java-sdk/src/main/java/com/microsoft/azure/toolkit/intellij/java/sdk/models/package-info.java
@@ -4,4 +4,4 @@
 /**
  * Package containing models used by the Azure SDK Build Tool.
  */
-package com.microsoft.azure.toolkit.intellij.java.sdk.report;
+package com.microsoft.azure.toolkit.intellij.java.sdk.models;
